fix: return the right drop target
